Rename quality_control submodules to levelN.py pattern
Align per-level files across diagnostics, qc_binaries, and quicklook
with the data_models/levelN.py naming:

  diagnostics/level_N.py -> levelN.py
  qc_binaries/checks_lN.py -> levelN.py
  quicklook/plot_lN.py -> levelN.py

Updates submodule imports in the package __init__ files plus tests
(test_qc, test_qlp_l0, test_qlp_l1), scripts/qlp.py, and the science
recipe. No logic changes.
