+++ kpfpipe/quality_control/diagnostics/__init__.py
@@ -1,6 +1,6 @@
 from kpfpipe.quality_control.diagnostics.base import Diagnostics
-from kpfpipe.quality_control.diagnostics.level_0 import DiagL0
-from kpfpipe.quality_control.diagnostics.level_1 import DiagL1
-from kpfpipe.quality_control.diagnostics.level_2 import DiagL2
+from kpfpipe.quality_control.diagnostics.level0 import DiagL0
+from kpfpipe.quality_control.diagnostics.level1 import DiagL1
+from kpfpipe.quality_control.diagnostics.level2 import DiagL2
 
 __all__ = ["Diagnostics", "DiagL0", "DiagL1", "DiagL2"]
